split MergingAndConvertingBlockInputStream
